+++ extensions/origins/source/aspects/amqp.js
@@ -1,0 +1,9 @@
+'use strict'
+
+/**
+ * @implements {toa.origins.amqp.Aspect}
+ */
+class Aspect {
+}
+
+exports.Aspect = Aspect
