+++ extensions/origins/source/aspects/amqp.test.js
@@ -1,0 +1,8 @@
+'use strict'
+
+const { Aspect } = require('./amqp')
+
+it('should be', async () => {
+  expect(Aspect).toBeInstanceOf(Function)
+})
+
