+++ extensions/origins/types/amqp.ts
@@ -1,0 +1,9 @@
+import * as _extensions from '@toa.io/core/types/extensions'
+
+declare namespace toa.origins.amqp {
+
+  interface Aspect extends _extensions.Aspect {
+    invoke(method: string, ...args: any[]): Promise<any>
+  }
+
+}
